+++ bin/phing-visualizer.php
@@ -5,11 +5,11 @@ try {
 
     $options = getopt('i:o:f:', ['input:', 'output:', 'format:']);
 
-    $input = $options['i'] ?? $options['input'] ?? null;
+    $input  = $options['i'] ?? $options['input'] ?? null;
     $output = $options['o'] ?? $options['output'] ?? null;
     $format = $options['f'] ?? $options['format'] ?? null;
 
-    $diagram = new Jawira\PhingVisualiser\Diagram($input);
+    $diagram = new Jawira\PhingVisualizer\Diagram($input);
     $diagram->save($format, $output);
 } catch (Exception $ex) {
     $msg = $ex->getMessage();